+++ src/providers/index.js
@@ -1,4 +1,5 @@
 module.exports = ngModule => {
+  require('./formlyTypes')(ngModule);
   require('./formlyUsability')(ngModule);
   require('./formlyConfig')(ngModule);
   require('./formlyVersion')(ngModule);