+++ webpack.config.js
@@ -51,7 +51,7 @@ var baseConfig = {
   module: {
     loaders: [
       {test: /\.html$/, loader: 'raw', exclude: exclude},
-      {test: /\.js$/, loader: ngAnnotateLoader + '!6to5!jshint', exclude: exclude},
+      {test: /\.js$/, loader: ngAnnotateLoader + '!babel!jshint', exclude: exclude},
       {test: /sinon.*\.js$/, loader: 'imports?define=>false'}
     ]
   }